myfastapi/pyproject.toml
[tool.poetry]
name = "myfastapi"
version = "0.1.0"
description = ""
authors = ["Marek <marecki.mail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.13.4"
fastapi = "^0.65.1"
SQLAlchemy = "^1.4.15"
PyJWT = "^2.1.0"
python-decouple = "^3.4"
pydantic = {extras = ["email"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
